% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{check_ecosystem}
\alias{check_ecosystem}
\title{Check input against possible ecosystems available}
\usage{
check_ecosystem(ecosystem, suppressMessages = TRUE)
}
\description{
Check input against possible ecosystems available
}
