% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_rosv}
\alias{copy_rosv}
\title{Create a copy of the {{rosv}} object}
\usage{
copy_rosv(x, ...)
}
\arguments{
\item{x}{Object to copy.}

\item{...}{Additional parameters sent to R6's clone method.}
}
\description{
Since R6 classes have reference semantics, to escape updating original objects,
a clone should be made.
}
\examples{
original_obj <- osvQuery1$new(name = 'readxl', ecosystem = 'CRAN'))
new_obj <- copy_rosv(original_obj)
}
