% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_osv_list}
\alias{create_osv_list}
\title{Create list of packages identified in OSV database}
\usage{
create_osv_list(
  type = "pypi",
  delim = "\\t",
  refresh = FALSE,
  clear_cache = FALSE
)
}
\arguments{
\item{type}{Character value of either 'pypi' or 'cran'.}

\item{delim}{The deliminator to separate the package and version details.}

\item{refresh}{Force refresh of the cache to grab latest details from OSV databases.}

\item{clear_cache}{Boolean value, to force clearing of the existing cache upon exiting function.}
}
\description{
Create list of packages identified in OSV database
}
\examples{
\dontrun{
pypi_vul <- create_osv_list()
writeLines(pypi_vul, 'pypi_vul.txt')

cran_vul <- create_osv_list(type = 'cran', delim = ',')
writeLines(cran_vul, 'cran_vul.csv')
}
}
