% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fetch_ecosystems}
\alias{fetch_ecosystems}
\title{Fetch all available ecosystems}
\usage{
fetch_ecosystems(offline = FALSE, refresh = FALSE)
}
\arguments{
\item{offline}{Boolean, determine if use list bundled with package.}

\item{refresh}{Boolean, force refresh of cache when using online list.}
}
\description{
Fetch all available ecosystems
}
