% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_pypi_pkg}
\alias{normalize_pypi_pkg}
\title{Normalize package name to PyPI expectation}
\usage{
normalize_pypi_pkg(pkg_name)
}
\arguments{
\item{pkg_name}{Vector of package names.}
}
\description{
Perform some formatting as PyPI is case insensitive and underscore, period, and hyphens
as long runs are not recognized (- is same as --).
}
