% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{osv_count_vulns}
\alias{osv_count_vulns}
\title{Count the number of reported vulnerabilities}
\usage{
osv_count_vulns(name, ecosystem, ...)
}
\arguments{
\item{name}{Character vector of package names.}

\item{ecosystem}{Character vector of ecosystem(s) within which the package(s) exist.}

\item{...}{Any other parameters to pass to nested functions.}
}
\value{
A named vector of numeric values indicating vulnerabilities.
}
\description{
Search the OSV database, by package name and its respective ecosystem, and count the number
of discovered vulnerabilities listed.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
osv_count_vulns(c('dask', 'dplyr'), c('PyPI', 'CRAN'))
\dontshow{\}) # examplesIf}
}
