% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{osv_query_1}
\alias{osv_query_1}
\title{Query OSV API for individual package vulnerabilities}
\usage{
osv_query_1(
  packages = NA,
  version = NA,
  ecosystem = NA,
  page_token = NA,
  body_only = TRUE,
  ...
)
}
\arguments{
\item{packages}{Name of package.}

\item{version}{Version of package.}

\item{ecosystem}{Ecosystem package lives within.}

\item{page_token}{When large number of results, next response to complete set requires a page_token.}

\item{body_only}{Boolean value to return entire response or just the body content.}

\item{...}{Additional parameters, for future development.}
}
\description{
Query OSV API for individual package vulnerabilities
}
\seealso{
\href{https://ossf.github.io/osv-schema/#affectedpackage-field}{Ecosystem list}
}
