% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{osv_querybatch}
\alias{osv_querybatch}
\title{Query OSV API for vulnerabilities given a vector of packages}
\usage{
osv_querybatch(
  packages = NA,
  version = NA,
  ecosystem = NA,
  page_token = NA,
  body_only = TRUE,
  ...
)
}
\arguments{
\item{packages}{Name of package.}

\item{version}{Version of package.}

\item{ecosystem}{Ecosystem package lives within.}

\item{page_token}{When large number of results, next response to complete set requires a page_token.}

\item{body_only}{Boolean value to return entire response or just the body content.}

\item{...}{Additional parameters, for future development.}
}
\description{
Each query needs to be constructed from the provided set of vectors. Default
will be \code{NA} and thereby empty/null in the JSON request. If some values in the vector
are missing, use NA For large queries, the conversion to a formatted JSON
request can be parallelized via {future}.
}
\details{
This returns the vulnerability ID and modified fields only, as per API instruction.
}
\seealso{
\href{https://ossf.github.io/osv-schema/#affectedpackage-field}{Ecosystem list}
}
