% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{osv_querybatch2}
\alias{osv_querybatch2}
\title{Query OSV API for vulnerabilities given a vector of packages}
\usage{
osv_querybatch2(cache = TRUE, ...)
}
\arguments{
\item{cache}{Boolean.}

\item{...}{Additional parameters, for future development.}
}
\value{
An R6 object containing API query contents.
}
\description{
Each query needs to be constructed from the provided set of vectors. Default
will be \code{NA} and thereby empty/null in the JSON request. If some values in the vector
are missing, use NA For large queries, the conversion to a formatted JSON
request can be parallelized via {future}.
}
\details{
This returns the vulnerability ID and modified fields only, as per API instruction.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
osv_querybatch(c("commonmark", "dask"), ecosystem = c('CRAN', 'PyPI'))
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://ossf.github.io/osv-schema/#affectedpackage-field}{Ecosystem list}
}
