% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scans.R
\name{osv_scan}
\alias{osv_scan}
\title{Use OSV database to scan for vulnerabilities}
\usage{
osv_scan(mode, ...)
}
\arguments{
\item{mode}{The kind of scan to perform.}

\item{...}{Parameters passed to specific underlying functions for mode selected.}
}
\value{
A data.frame specifying which packages are vulnerable or not.
}
\description{
Scan project based upon specified mode to determine if any vulnerable packages are detected.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
osv_scan('r_libpath')
\dontshow{\}) # examplesIf}
}
