% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosv-package.R
\docType{package}
\name{rosv-package}
\alias{rosv}
\alias{rosv-package}
\title{rosv: Client to Access and Operate on the 'Open Source Vulnerability' API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Connect, query, and operate on information available from the 'Open Source Vulnerability' database \url{https://osv.dev/}. Although 'CRAN' has vulnerabilities listed, these are few compared to projects such as 'PyPI'. With tighter integration between 'R' and 'Python', having an 'R' specific package to access details about vulnerabilities from various sources is a worthwhile enterprise.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://al-obrien.github.io/rosv/}
  \item \url{https://github.com/al-obrien/rosv}
  \item Report bugs at \url{https://github.com/al-obrien/rosv/issues}
}

}
\author{
\strong{Maintainer}: Allen OBrien \email{allen.g.obrien@gmail.com} [copyright holder]

}
\keyword{internal}
