% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_query.R
\name{RosvQuery1}
\alias{RosvQuery1}
\title{R6 Class for OSV Query Endpoint}
\description{
Pageination not implemetned yet, waiting for httr2 updates to add, then will handle automatically
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RosvQuery1-run}{\code{RosvQuery1$run()}}
\item \href{#method-RosvQuery1-print}{\code{RosvQuery1$print()}}
\item \href{#method-RosvQuery1-clone}{\code{RosvQuery1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQuery1-run"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQuery1-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQuery1$run(
  commit = NULL,
  version = NULL,
  name = NULL,
  ecosystem = NULL,
  purl = NULL,
  page_token = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQuery1-print"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQuery1-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQuery1$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQuery1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQuery1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQuery1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
