% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_query.R
\name{RosvQueryBatch}
\alias{RosvQueryBatch}
\title{R6 Class for OSV Querybatch Endpoint}
\description{
Batches are enforced to only process by commit hash, purl, or name+ecosystem.
This avoids some confusion as to which is taken preferentially and simplifies query creation.
}
\section{Super class}{
\code{rosv::RosvQuery1} -> \code{RosvQueryBatch}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RosvQueryBatch-run}{\code{RosvQueryBatch$run()}}
\item \href{#method-RosvQueryBatch-clone}{\code{RosvQueryBatch$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rosv" data-topic="RosvQuery1" data-id="print"><a href='../../rosv/html/RosvQuery1.html#method-RosvQuery1-print'><code>rosv::RosvQuery1$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQueryBatch-run"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQueryBatch-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQueryBatch$run(
  commit = NULL,
  version = NULL,
  name = NULL,
  ecosystem = NULL,
  purl = NULL,
  page_token = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQueryBatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQueryBatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQueryBatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
