% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_rosv}
\alias{validate_rosv}
\title{Validate if object is made by {{rosv}}}
\usage{
validate_rosv(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Invisibly returns TRUE if validation is successful, otherwise will error.
}
\description{
Validate if object is made by {{rosv}}
}
\examples{
# example code
validate_rosv(RosvQuery1$new(name = 'readxl', ecosystem = 'CRAN'))
}
