% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ecosystem}
\alias{check_ecosystem}
\title{Check input against possible ecosystems available}
\usage{
check_ecosystem(ecosystem, suppressMessages = TRUE)
}
\arguments{
\item{ecosystem}{Character value for ecosystem(s) to check.}

\item{suppressMessages}{Boolean value whether or not to suppress any messages.}
}
\value{
A character vector of the same input if all are valid ecosystem names.
}
\description{
Ensures that inputs for ecosystem are valid based upon what is available in OSV database.
}
\details{
Will attempt to grab latest file and cache for the session. If cannot access
the online version, will use a local copy that is shipped with the package.
}
\examples{
# Passes
rosv:::check_ecosystem(c('PyPI', 'CRAN'))

# Fails
try(rosv:::check_ecosystem(c('notvalid', 'pypi')))

}
\seealso{
\code{\link{check_ecosystem}}
}
