% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ecosystem}
\alias{check_ecosystem}
\title{Check input against possible ecosystems available}
\usage{
check_ecosystem(ecosystem, suppressMessages = TRUE)
}
\arguments{
\item{ecosystem}{Character value for ecosystem(s) to check.}

\item{suppressMessages}{Boolean value whether or not to suppress any messages.}
}
\value{
A character vector, the same as input if all are valid ecosystem names.
}
\description{
Internal function that ensures inputs for ecosystem are valid based upon what
is available in the OSV database.
}
\details{
Will attempt to grab latest file and cache for the current R session. If session
cannot access the online version, it will use a local copy shipped with the package.
}
\seealso{
\code{\link{fetch_ecosystems}}
}
