% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_osv_cache}
\alias{clear_osv_cache}
\title{Reset cached results of OSV calls}
\usage{
clear_osv_cache()
}
\value{
Invisibly returns a logical value of \code{TRUE} if cache cleared without error.
}
\description{
A thin wrapper around \code{\link[memoise]{forget}} to clear cached results.
}
\examples{
clear_osv_cache()
}
