% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ppm_blacklist}
\alias{create_ppm_blacklist}
\title{Create blacklist commands for Posit Package Manager from OSV data}
\usage{
create_ppm_blacklist(osv_list, delim, flags = NULL)
}
\arguments{
\item{osv_list}{Output from \code{create_osv_list()}.}

\item{delim}{The delimiter used when creating \code{osv_list}.}

\item{flags}{Global flag to apply to the rspm commands.}
}
\description{
Create blacklist commands for Posit Package Manager from OSV data
}
\examples{
\dontrun{
pypi_vul <- create_osv_list(delim = ',')
cmd_blist <- create_ppm_blacklist(pypi_vul, delim = ',', flags = '--source=pypi')
}
}
