% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ppm_xref_whitelist}
\alias{create_ppm_xref_whitelist}
\title{Cross reference a whitelist of packages to a vulnerability database}
\usage{
create_ppm_xref_whitelist(
  packages,
  osv_list,
  type = "pypi",
  version_placeholder = " "
)
}
\arguments{
\item{packages}{Character vector of package names.}

\item{osv_list}{OSV data/list created from \code{create_osv_list}.}

\item{type}{Determine what type of OSV list is being used (currently only works with pypi).}

\item{version_placeholder}{Value used when creating the \code{osv_list} from \code{create_osv_list}.}
}
\description{
Cross reference a whitelist of packages to a vulnerability database
}
\examples{
\dontrun{
python_pkg <- c('dask', 'tensorflow', 'keras')
pypi_vul <- create_osv_list(as.data.frame = TRUE)
xref_pkg_list <- create_ppm_xref_whitelist(python_pkg, pypi_vul)
writeLines(xref_pkg_list, 'requirements.txt')
}
}
\seealso{
\href{https://packaging.python.org/en/latest/specifications/name-normalization/}{PyPI package normalization}
}
