% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_osv}
\alias{download_osv}
\title{Download helper for OSV data}
\usage{
download_osv(type = "pypi", refresh = FALSE)
}
\arguments{
\item{type}{Character value of either 'pypi' or 'cran'.}

\item{refresh}{Force refresh of the cache to grab latest details from OSV databases.}
}
\description{
Helper function to assist in downloading vulnerabilities information from OSV database.
}
