% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_vul_info}
\alias{extract_vul_info}
\title{Extract key OSV information from JSON}
\usage{
extract_vul_info(file, delim = "\\t", version_placeholder = " ")
}
\arguments{
\item{file}{File path to the folder containing the OSV JSON files.}

\item{delim}{The deliminator to separate the package and version details.}

\item{version_placeholder}{Value to fill if no versions are listed for package.}
}
\description{
Use the downloaded JSON dataset from OSV and extract key details on package and versions listed.
}
