% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_ecosystems}
\alias{fetch_ecosystems}
\title{Fetch all available ecosystems}
\usage{
fetch_ecosystems(offline = FALSE, refresh = FALSE)
}
\arguments{
\item{offline}{Boolean, determine if using list bundled with package.}

\item{refresh}{Boolean, force refresh of cache when using online list.}
}
\value{
A data.frame containing all the ecosystem names available in the OSV database.
}
\description{
Internal function used to fetch the available ecosystems in the OSV API.
}
\details{
The \code{refresh} parameter can be used to force the data to be pulled again
even if one is available in the cached location. Since a fresh pull is performed
for each R session, it is unlikely that this parameter is required and is primarily
reserved for future use if functionality necessitates.
}
\seealso{
\code{\link{check_ecosystem}}
}
