% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rosv}
\alias{get_rosv}
\title{Internal function to assist with extracting details fro {{rosv}} objects}
\usage{
get_rosv(x, field)
}
\description{
Internal function to assist with extracting details fro {{rosv}} objects
}
