% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_pypi_pkg}
\alias{normalize_pypi_pkg}
\title{Normalize package name to PyPI expectation}
\usage{
normalize_pypi_pkg(pkg_name)
}
\arguments{
\item{pkg_name}{Character vector of package names.}
}
\value{
Character vector of normalized PyPI package names
}
\description{
Perform package name formatting as PyPI is case insensitive and long runs
of underscore, period, and hyphens are not recognized (- is same as --).
}
\examples{
normalize_pypi_pkg(c('Dask', 'TenSorFlow'))

}
\seealso{
\href{https://packaging.python.org/en/latest/specifications/name-normalization/}{PyPI Package Normalization}
}
