% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{osv_query_1}
\alias{osv_query_1}
\title{Query OSV API for individual package vulnerabilities}
\usage{
osv_query_1(
  name = NULL,
  version = NULL,
  ecosystem = NULL,
  commit = NULL,
  purl = NULL,
  page_token = NA,
  ...
)
}
\arguments{
\item{name}{Name of package.}

\item{version}{Version of package.}

\item{ecosystem}{Ecosystem package lives within (must be set if using \code{name}).}

\item{commit}{Commit hash to query against (do not use when version set).}

\item{purl}{URL for package (do not use if name or ecosystem set).}

\item{page_token}{When large number of results, next response to complete set requires a page_token.}

\item{...}{Additional parameters, for future development.}
}
\value{
A list containing API query contents.
}
\description{
Query OSV API for individual package vulnerabilities
}
\examples{
osv_query_1(commit = '6879efc2c1596d11a6a6ad296f80063b558d5e0f')


}
\seealso{
\href{https://ossf.github.io/osv-schema/#affectedpackage-field}{Ecosystem list}
}
