% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{osv_querybatch}
\alias{osv_querybatch}
\alias{.osv_querybatch}
\alias{.osv_querybatch_cache}
\title{Query OSV API for vulnerabilities given a vector of packages}
\usage{
osv_querybatch(
  name = NULL,
  version = NULL,
  ecosystem = NULL,
  commit = NULL,
  purl = NULL,
  page_token = NA,
  parse = TRUE,
  cache = TRUE,
  ...
)

.osv_querybatch(
  name = NULL,
  version = NULL,
  ecosystem = NULL,
  commit = NULL,
  purl = NULL,
  page_token = NA,
  parse = TRUE,
  cache = TRUE,
  ...
)

.osv_querybatch_cache(
  name = NULL,
  version = NULL,
  ecosystem = NULL,
  commit = NULL,
  purl = NULL,
  page_token = NA,
  parse = TRUE,
  cache = TRUE,
  ...
)
}
\arguments{
\item{name}{Name of package.}

\item{version}{Version of package.}

\item{ecosystem}{Ecosystem package lives within (must be set if using \code{name}).}

\item{commit}{Commit hash to query against (do not use when version set).}

\item{purl}{URL for package (do not use if name or ecosystem set).}

\item{page_token}{When large number of results, next response to complete set requires a page_token.}

\item{parse}{Boolean value to set if the content field should be parsed from JSON list format.}

\item{cache}{Boolean value to determine if should use a cached version of the function and API results.}

\item{...}{Additional parameters passed to nested functions.}
}
\value{
An R6 object containing API query contents.
}
\description{
Each query needs to be constructed from the provided set of vectors. Default
will be \code{NA} and thereby empty/null in the JSON request. If some values in the vector
are missing, use NA For large queries, the conversion to a formatted JSON
request can be parallelized via {future}.
}
\details{
This returns the vulnerability ID and modified fields only, as per API instruction.
}
\section{Functions}{
\itemize{
\item \code{.osv_querybatch()}: Internal function to run osv_querybatch without caching

\item \code{.osv_querybatch_cache()}: Internal function to run a memoise and cached version of osv_querybatch

}}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
osv_querybatch(c("commonmark", "dask"), ecosystem = c('CRAN', 'PyPI'))
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://ossf.github.io/osv-schema/#affectedpackage-field}{Ecosystem list}
}
