% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{osv_vulns}
\alias{osv_vulns}
\title{Query OSV API for vulnerabilities based on ID}
\usage{
osv_vulns(vulns_ids, body_only = TRUE)
}
\arguments{
\item{vulns_ids}{Vector of vulnerability IDs.}

\item{body_only}{Boolean value to return entire response or just the body content.

Is usually paired with the batch outputs to grab more specific information.}
}
\description{
Query OSV API for vulnerabilities based on ID
}
