% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{osv_vulns}
\alias{osv_vulns}
\title{Query OSV API for vulnerabilities based on ID}
\usage{
osv_vulns(vuln_ids)
}
\arguments{
\item{vuln_ids}{Vector of vulnerability IDs.

Is usually paired with the batch outputs to grab more specific information.}
}
\description{
Query OSV API for vulnerabilities based on ID
}
