% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{osv_vulns}
\alias{osv_vulns}
\title{Query OSV API for vulnerabilities based on ID}
\usage{
osv_vulns(vuln_ids, parse = TRUE)
}
\arguments{
\item{vuln_ids}{Vector of vulnerability IDs.}

\item{parse}{Boolean value to set if the content field should be parsed from JSON list format.}
}
\value{
An R6 object containing API query contents.
}
\description{
Use vulnerability IDs to extract more details information. Usually is paired with \code{osv_querybatch}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
vulns <- osv_vulns("RSEC-2023-8")
get_content(vulns)
\dontshow{\}) # examplesIf}
}
