% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_query.R
\name{RosvQuery1}
\alias{RosvQuery1}
\title{R6 Class for OSV Query Endpoint}
\description{
An R6 class to provide a lower-level interface to the query endpoint of the OSV API.
}
\details{
Pageination not implemented yet, waiting for httr2 updates to add, then will handle automatically
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{request}}{Request object made by \code{httr2}.}

\item{\code{content}}{Body contents of response from OSV API.}

\item{\code{response}}{Response object returned from OSV API.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RosvQuery1-run}{\code{RosvQuery1$run()}}
\item \href{#method-RosvQuery1-print}{\code{RosvQuery1$print()}}
\item \href{#method-RosvQuery1-clone}{\code{RosvQuery1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQuery1-run"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQuery1-run}{}}}
\subsection{Method \code{run()}}{
Perform the request and return response for OSV API call.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQuery1$run(
  commit = NULL,
  version = NULL,
  name = NULL,
  ecosystem = NULL,
  purl = NULL,
  page_token = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commit}}{Commit hash to query against (do not use when version set).}

\item{\code{version}}{Version of package.}

\item{\code{name}}{Name of package.}

\item{\code{ecosystem}}{Ecosystem package lives within (must be set if using \code{name}).}

\item{\code{purl}}{URL for package (do not use if name or ecosystem set).}

\item{\code{page_token}}{When large number of results, next response to complete set requires a page_token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQuery1-print"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQuery1-print}{}}}
\subsection{Method \code{print()}}{
Print basic details of query object to screen.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQuery1$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Reserved for possible future use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQuery1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQuery1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQuery1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
