% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_query.R
\name{RosvQueryBatch}
\alias{RosvQueryBatch}
\title{R6 Class for OSV Querybatch Endpoint}
\description{
An R6 class to provide a lower-level interface to the querybatch
endpoint of the OSV API. Batches are enforced to only process by commit hash, purl, or name+ecosystem.
This avoids some confusion as to which is taken preferentially and simplifies query creation.
}
\section{Super class}{
\code{\link[rosv:RosvQuery1]{rosv::RosvQuery1}} -> \code{RosvQueryBatch}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RosvQueryBatch-new}{\code{RosvQueryBatch$new()}}
\item \href{#method-RosvQueryBatch-run}{\code{RosvQueryBatch$run()}}
\item \href{#method-RosvQueryBatch-clone}{\code{RosvQueryBatch$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rosv" data-topic="RosvQuery1" data-id="print"><a href='../../rosv/html/RosvQuery1.html#method-RosvQuery1-print'><code>rosv::RosvQuery1$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQueryBatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQueryBatch-new}{}}}
\subsection{Method \code{new()}}{
Set the core request details for subsequent use when called in \code{run()} method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQueryBatch$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQueryBatch-run"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQueryBatch-run}{}}}
\subsection{Method \code{run()}}{
Perform the request and return response for OSV API call.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQueryBatch$run(
  commit = NULL,
  version = NULL,
  name = NULL,
  ecosystem = NULL,
  purl = NULL,
  page_token = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commit}}{Commit hash to query against (do not use when version set).}

\item{\code{version}}{Version of package.}

\item{\code{name}}{Name of package.}

\item{\code{ecosystem}}{Ecosystem package lives within (must be set if using \code{name}).}

\item{\code{purl}}{URL for package (do not use if name or ecosystem set).}

\item{\code{page_token}}{When large number of results, next response to complete set requires a page_token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvQueryBatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RosvQueryBatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvQueryBatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
