% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_query.R
\name{RosvVulns}
\alias{RosvVulns}
\title{R6 Class for OSV Vulns Endpoint}
\description{
R6 Class for OSV Vulns Endpoint

R6 Class for OSV Vulns Endpoint
}
\section{Super class}{
\code{rosv::RosvQuery1} -> \code{RosvVulns}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RosvVulns-run}{\code{RosvVulns$run()}}
\item \href{#method-RosvVulns-print}{\code{RosvVulns$print()}}
\item \href{#method-RosvVulns-clone}{\code{RosvVulns$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvVulns-run"></a>}}
\if{latex}{\out{\hypertarget{method-RosvVulns-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvVulns$run(vuln_ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvVulns-print"></a>}}
\if{latex}{\out{\hypertarget{method-RosvVulns-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvVulns$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RosvVulns-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RosvVulns-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RosvVulns$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
